markers (and tick directions)
