+++ tests/test_marker.py
@@ -1,13 +1,17 @@
 """Tests for marker styles."""
 
-from tueplots import markers
 import matplotlib.pyplot as plt
 
-def test_dark_edge():
-    plt.rcParams.update(markers.dark_edge())
+from tueplots import marker
+
+
+def test_with_edge():
+    plt.rcParams.update(marker.with_edge())
+
 
 def test_inverted():
-    plt.rcParams.update(markers.inverted())
+    plt.rcParams.update(marker.inverted())
+
 
 def test_cycler():
-    plt.rcParams.update(markers.cycler())
+    plt.rcParams.update(marker.cycler())
